Relocate minicanvas component to its own file
