+++ src/components/Dashboard/ComplexWidget/MiniCanvas.jsx
@@ -1,7 +1,7 @@
-import React, { Component } from "react";
-import { getWidgetDefinition } from "../widgets/widgetDefinitions";
-import PropTypes from "prop-types";
-import { widget } from "../../../propTypes/propTypes";
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { getWidgetDefinition } from '../widgets/widgetDefinitions';
+import { widget } from '../../../propTypes/propTypes';
 
 class MiniCanvas extends Component {
   componentForWidget(widget) {
@@ -9,18 +9,18 @@ class MiniCanvas extends Component {
   }
 
   deviceForWidget(widget) {
-    return widget.device === "__parent__" ? this.props.device : widget.device;
+    return widget.device === '__parent__' ? this.props.device : widget.device;
   }
 
   valueAndTimeForWidget(widget) {
-    if (this.props.mode !== "run") {
+    if (this.props.mode !== 'run') {
       return {};
     }
 
     const device = this.deviceForWidget(widget);
     const attribute = widget.attribute;
     const key = `${device}/${attribute}`;
-    return this.props.attributes[key] || {};
+    return this.props.attributes[key] || {};
   }
 
   render() {
@@ -32,38 +32,38 @@ class MiniCanvas extends Component {
     const minX = widgets.map(({ x }) => x).reduce((a, b) => Math.min(a, b));
     const minY = widgets.map(({ y }) => y).reduce((a, b) => Math.min(a, b));
 
-    const isLibrary = this.props.mode === "library";
+    const isLibrary = this.props.mode === 'library';
 
     const style = isLibrary
       ? {
-          position: "relative",
-          height: "100px",
-          overflow: "hidden",
+          position: 'relative',
+          height: '100px',
+          overflow: 'hidden',
           background: `repeating-linear-gradient(
-            rgb(232, 239, 249),
-            rgb(232, 239, 249) 1px,
-            rgba(0, 0, 0, 0) 1px,
-            rgba(0, 0, 0, 0) 15px
-          ),
-          repeating-linear-gradient(
-            90deg,
-            rgb(232, 239, 249),
-            rgb(232, 239, 249) 1px,
-            rgba(0, 0, 0, 0) 1px,
-            rgba(0, 0, 0, 0) 15px
-          )`
+              rgb(232, 239, 249),
+              rgb(232, 239, 249) 1px,
+              rgba(0, 0, 0, 0) 1px,
+              rgba(0, 0, 0, 0) 15px
+            ),
+            repeating-linear-gradient(
+              90deg,
+              rgb(232, 239, 249),
+              rgb(232, 239, 249) 1px,
+              rgba(0, 0, 0, 0) 1px,
+              rgba(0, 0, 0, 0) 15px
+            )`
         }
-      : { width: "300px", height: "200px" };
+      : { width: '300px', height: '200px' };
 
     const FadeOut = () => (
       <div
         style={{
-          position: "absolute",
+          position: 'absolute',
           top: 0,
           left: 0,
-          width: "100%",
-          height: "100%",
-          boxShadow: "inset -10px -10px 20px 0px white",
+          width: '100%',
+          height: '100%',
+          boxShadow: 'inset -10px -10px 20px 0px white',
           zIndex: 10000
         }}
       />
@@ -82,9 +82,9 @@ class MiniCanvas extends Component {
             return (
               <div
                 key={i}
-                className={"Widget"}
+                className={'Widget'}
                 style={{
-                  position: "absolute",
+                  position: 'absolute',
                   left: x - minX,
                   top: y - minY
                 }}
@@ -105,34 +105,10 @@ class MiniCanvas extends Component {
     );
   }
 }
+
 MiniCanvas.propTypes = {
   mode: PropTypes.string,
   widgets: PropTypes.arrayOf(widget)
 };
 
-function complexWidgetComponent(canvas) {
-  return class ComplexWidget extends Component {
-    render() {
-      return (
-        <MiniCanvas
-          device={this.props.device}
-          widgets={canvas.widgets}
-          attributes={this.props.attributes}
-          mode={this.props.mode}
-        />
-      );
-    }
-  };
-}
-
-export function complexWidgetDefinition(canvas) {
-  const { id, name } = canvas;
-  return {
-    type: `CANVAS_${id}`,
-    name,
-    component: complexWidgetComponent(canvas),
-    fields: ["device"],
-    params: [],
-    __canvas__: id
-  };
-}
+export default MiniCanvas;
